Add Dockerfile and nginx config for testing site
This is needed, because mkdocs defaults to using directory URLs, and
cannot function the way GitHub Pages handles pages by default.

The Docker image built from this will emulate the behaviour of GitHub
Pages's pretty URL setup, which makes the `.html` ending not necessary.
We rely on this behaviour treewide, and mkdocs doesn't provide that for
us.